Add neccesary changes to support the new alastria name service contract
