+++ package.json
@@ -9,7 +9,7 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "postinstall": "run-s build",
-    "compile": "git clone --branch develop https://github.com/alastria/temp-alastriaID-truffle-contracts.git && npm install --prefix temp-alastriaID-truffle-contracts && npm run compile --prefix temp-alastriaID-truffle-contracts && node src/configFile.js && tsc",
+    "compile": "git clone --branch deploy/AlastriaIdentityEntity https://github.com/alastria/temp-alastriaID-truffle-contracts.git && npm install --prefix temp-alastriaID-truffle-contracts && npm run compile --prefix temp-alastriaID-truffle-contracts && node src/configFile.js && tsc",
     "clean": "rimraf dist temp-alastriaID-truffle-contracts",
     "build": "run-s clean compile",
     "test": "mocha -r ts-node/register test/**/*.ts",
