+++ src/configFile.js
@@ -33,6 +33,7 @@ const config = {
   alastriaCredentialRegistry: contractsInfo.AlastriaCredentialRegistry,
   alastriaPresentationRegistry: contractsInfo.AlastriaPresentationRegistry,
   alastriaPublicKeyRegistry: contractsInfo.AlastriaPublicKeyRegistry,
+  alastriaNameService: contractsInfo.AlastriaNameService,
   basicTransaction: {
     to: '0x0000000000000000000000000000000000000000',
     data: '0x0',
