+++ src/txFactory/identityManagerTransactionFactory.ts
@@ -203,166 +203,6 @@ export function isIdentityIssuer(web3, didIssuer) {
   return transaction
 }
 
-/**
- * function addEntity(address _addressEntity, string _name, string _cif, string _url_logo, string _url_createAID, string _url_AOA, bool _active) public onlyIdentityEntity(msg.sender)
- * @param web3
- * @param didEntity
- * @param name
- * @param cif
- * @param urlLogo
- * @param urlCreateAID
- * @param urlAOA
- * @param status
- */
-export function addEntity(
-  web3,
-  didEntity,
-  name,
-  cif,
-  urlLogo,
-  urlCreateAID,
-  urlAOA,
-  status
-) {
-  const entityAddr = AIdUtils.getProxyAddress(didEntity)
-  const transaction = Object.assign({}, config.basicTransaction)
-  const delegatedData = web3.eth.abi.encodeFunctionCall(
-    config.contractsAbi.AlastriaIdentityEntity.addEntity,
-    [entityAddr, name, cif, urlLogo, urlCreateAID, urlAOA, status]
-  )
-  transaction.data = delegated(web3, delegatedData)
-  transaction.to = config.alastriaIdentityManager
-  transaction.gasLimit = 600000
-  return transaction
-}
-
-/**
- * function setNameEntity(address _addressEntity, string _name) public
- * @param web3
- * @param didEntity
- * @param name
- */
-export function setNameEntity(web3, didEntity, name) {
-  const entityAddr = AIdUtils.getProxyAddress(didEntity)
-  const transaction = Object.assign({}, config.basicTransaction)
-  const delegatedData = web3.eth.abi.encodeFunctionCall(
-    config.contractsAbi.AlastriaIdentityEntity.setNameEntity,
-    [entityAddr, name]
-  )
-  transaction.data = delegated(web3, delegatedData)
-  transaction.to = config.alastriaIdentityManager
-  transaction.gasLimit = 600000
-  return transaction
-}
-
-/**
- * function setCifEntity(address _addressEntity, string _cif) public
- * @param web3
- * @param didEntity
- * @param cif
- */
-export function setCifEntity(web3, didEntity, cif) {
-  const entityAddr = AIdUtils.getProxyAddress(didEntity)
-  const transaction = Object.assign({}, config.basicTransaction)
-  const delegatedData = web3.eth.abi.encodeFunctionCall(
-    config.contractsAbi.AlastriaIdentityEntity.setCifEntity,
-    [entityAddr, cif]
-  )
-  transaction.data = delegated(web3, delegatedData)
-  transaction.to = config.alastriaIdentityManager
-  transaction.gasLimit = 600000
-  return transaction
-}
-
-/**
- * function setUrlLogo(address _addressEntity, string _url_logo) public
- * @param web3
- * @param didEntity
- * @param urlLogo
- */
-export function setUrlLogo(web3, didEntity, urlLogo) {
-  const entityAddr = AIdUtils.getProxyAddress(didEntity)
-  const transaction = Object.assign({}, config.basicTransaction)
-  const delegatedData = web3.eth.abi.encodeFunctionCall(
-    config.contractsAbi.AlastriaIdentityEntity.setUrlLogo,
-    [entityAddr, urlLogo]
-  )
-  transaction.data = delegated(web3, delegatedData)
-  transaction.to = config.alastriaIdentityManager
-  transaction.gasLimit = 600000
-  return transaction
-}
-
-/**
- * function setUrlCreateAID(address _addressEntity, string _url_createAID) public
- * @param web3
- * @param didEntity
- * @param urlCreateAID
- */
-export function setUrlCreateAID(web3, didEntity, urlCreateAID) {
-  const entityAddr = AIdUtils.getProxyAddress(didEntity)
-  const transaction = Object.assign({}, config.basicTransaction)
-  const delegatedData = web3.eth.abi.encodeFunctionCall(
-    config.contractsAbi.AlastriaIdentityEntity.setUrlCreateAID,
-    [entityAddr, urlCreateAID]
-  )
-  transaction.data = delegated(web3, delegatedData)
-  transaction.to = config.alastriaIdentityManager
-  transaction.gasLimit = 600000
-  return transaction
-}
-
-/**
- * function setUrlAOA(address _addressEntity, string _url_AOA) public
- * @param web3
- * @param didEntity
- * @param urlAOA
- */
-export function setUrlAOA(web3, didEntity, urlAOA) {
-  const entityAddr = AIdUtils.getProxyAddress(didEntity)
-  const transaction = Object.assign({}, config.basicTransaction)
-  const delegatedData = web3.eth.abi.encodeFunctionCall(
-    config.contractsAbi.AlastriaIdentityEntity.setUrlAOA,
-    [entityAddr, urlAOA]
-  )
-  transaction.data = delegated(web3, delegatedData)
-  transaction.to = config.alastriaIdentityManager
-  transaction.gasLimit = 600000
-  return transaction
-}
-
-/**
- * function getEntity(address _addressEntity) public view returns(string _name, string _cif, string _url_logo, string _url_createAID, string _url_AOA, bool _active)
- * @param web3
- * @param didEntity
- */
-export function getEntity(web3, didEntity) {
-  const entityAddr = AIdUtils.getProxyAddress(didEntity)
-  const transaction = Object.assign({}, config.basicTransaction)
-  transaction.data = web3.eth.abi.encodeFunctionCall(
-    config.contractsAbi.AlastriaIdentityEntity.getEntity,
-    [entityAddr]
-  )
-  transaction.to = config.alastriaIdentityManager
-  transaction.gasLimit = 600000
-  return transaction
-}
-
-/**
- * function entitiesList() public view returns(address[])
- * @param web3
- */
-export function entitiesList(web3) {
-  const transaction = Object.assign({}, config.basicTransaction)
-  transaction.data = web3.eth.abi.encodeFunctionCall(
-    config.contractsAbi.AlastriaIdentityEntity.entitiesList,
-    []
-  )
-  transaction.to = config.alastriaIdentityManager
-  transaction.gasLimit = 600000
-  return transaction
-}
-
 function delegated(web3, delegatedData) {
   return web3.eth.abi.encodeFunctionCall(
     config.contractsAbi.AlastriaIdentityManager.delegateCall,
